+++ SunacCoordination/SunacCoordination/WebIO/cpp_base64/src/Base64Encoder.cpp
@@ -4,7 +4,7 @@ Base64Encoder::Base64Encoder() { base64_init_encodestate(&state); }
 std::string Base64Encoder::encode(const std::string &text)
 {
 	char *buffer = new char[text.length() * 2];
-	int numberOfBytesEncoded = base64_encode_block(text.c_str(), text.length(), buffer, &state);
+	int numberOfBytesEncoded = base64_encode_block(text.c_str(), (int)text.length(), buffer, &state);
 	numberOfBytesEncoded += base64_encode_blockend(buffer + numberOfBytesEncoded, &state);
 	buffer[numberOfBytesEncoded] = '\0';
 	std::string result(buffer);
